Update contactInfo.html
Just fill in the details whenever you get the chance, I think she wants a contact form too, so if it's possible to make one at the bottom without messing up the format, you should do that. If not we can make this an "about us" page and make another "contact" page with the form.